Extract coerceToForm utility
